% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breaks.R
\name{breaks-class}
\alias{breaks-class}
\alias{format.breaks}
\alias{print.breaks}
\alias{is.breaks}
\title{Class representing a set of intervals}
\usage{
\method{format}{breaks}(x, ...)

\method{print}{breaks}(x, ...)

is.breaks(x, ...)
}
\arguments{
\item{x}{A breaks object}

\item{...}{Unused}
}
\description{
Class representing a set of intervals
}
