% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breaks.R
\name{brk-left-right}
\alias{brk-left-right}
\alias{brk_right}
\alias{brk_left}
\title{Left- or right-closed breaks}
\usage{
brk_right(breaks, close_leftmost = TRUE)

brk_left(breaks, close_rightmost = TRUE)
}
\arguments{
\item{breaks}{A numeric vector which must be sorted.}

\item{close_leftmost, close_rightmost}{Logical: close the left/rightmost endpoint?}
}
\value{
A (function returning an) object of class \code{breaks}.
}
\description{
Left- or right-closed breaks
}
\examples{
chop(1:3, brk_left(1:3))
chop(1:3, brk_right(1:3))
chop(1:3, brk_left(1:3, FALSE))
}
