% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breaks.R
\name{brk_default}
\alias{brk_default}
\title{Create a standard set of breaks}
\usage{
brk_default(breaks)
}
\arguments{
\item{breaks}{A numeric vector.}
}
\value{
A (function which returns an) object of class \code{breaks}.
}
\description{
Create a standard set of breaks
}
\examples{

chop(1:10, c(2, 5, 8))
chop(1:10, brk_default(c(2, 5, 8)))

}
