% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breaks.R
\name{brk_fn}
\alias{brk_fn}
\title{Create breaks using an existing function}
\usage{
brk_fn(fn, ...)
}
\arguments{
\item{fn}{A function which returns a numeric vector of breaks.}

\item{...}{Further arguments to \code{fn}}
}
\value{
A function which returns an object of class \code{breaks}.
}
\description{
\code{brk_fn()} is a convenience wrapper: \code{chop(x, brk_fn(foo, ...))}
is the same as \code{chop(x, foo(x, ...))}.
}
\examples{

if (requireNamespace("scales")) {
  chop(rlnorm(10), brk_fn(scales::breaks_log(5)))
  # same as
  # x <- rlnorm(10)
  # chop(x, scales::breaks_log(5)(x))
}

}
