% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breaks.R
\name{brk_n}
\alias{brk_n}
\title{Breaks with a fixed number of elements}
\usage{
brk_n(n)
}
\arguments{
\item{n}{Integer: number of elements in each interval.}
}
\value{
A (function which returns an) object of class \code{breaks}.
}
\description{
\code{brk_n()} creates intervals containing a fixed number of elements. One
interval may have fewer elements.
}
\examples{
tab(runif(10), brk_n(2))
}
