% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breaks.R
\name{brk_quantiles}
\alias{brk_quantiles}
\title{Breaks using quantiles}
\usage{
brk_quantiles(probs, ...)
}
\arguments{
\item{probs}{Vector of probabilities for quantiles}

\item{...}{Arguments passed to \code{\link[=quantile]{quantile()}}}
}
\value{
A (function which returns an) object of class \code{breaks}.
}
\description{
Breaks using quantiles
}
\examples{
chop(c(1, 1, 2, 2, 3, 4), brk_quantiles(1:3/4))
}
