% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breaks.R
\name{brk_size}
\alias{brk_size}
\title{Fixed-size breaks}
\usage{
brk_size(size)
}
\arguments{
\item{size}{Integer size of breaks}
}
\value{
A (function returning an) object of class \code{breaks}.
}
\description{
\code{brk_size()} creates intervals containing a fixed number of (non-NA)
elements. One interval may have fewer elements.
}
\examples{
chop(runif(10), brk_size(2))
}
