% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breaks.R
\name{brk_width}
\alias{brk_width}
\title{Equal-width breaks}
\usage{
brk_width(width, start)
}
\arguments{
\item{width}{Width of intervals.}

\item{start}{Leftpoint of first interval. By default the lowest finite \code{x}.}
}
\value{
A (function which returns an) object of class \code{breaks}.
}
\description{
\code{brk_width()} creates breaks of equal width.
}
\examples{
chop(runif(10), brk_width(0.2, 0))
}
