% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chop.R
\name{chop.Date}
\alias{chop.Date}
\alias{chop.POSIXt}
\title{Chop Date objects}
\usage{
\method{chop}{Date}(x, breaks, labels = lbl_date(), ...)

\method{chop}{POSIXt}(x, breaks, labels = lbl_date("\%F \%T"), ...)
}
\arguments{
\item{x}{Vector of data to chop.}

\item{breaks}{Passed to \code{\link[=chop]{chop()}}.}

\item{labels}{Passed to \code{\link[=chop]{chop()}}.}

\item{...}{Passed to \code{\link[=chop]{chop()}}.}
}
\value{
A \link{factor} of the same length as \code{x}, representing the intervals containing
the value of \code{x}.
}
\description{
These functions chop Date and POSIXct objects. The default labels are given by
\code{\link[=lbl_date]{lbl_date()}}. \code{breaks} may be specified as \link{Date} or \link{POSIXt} objects, or
using any \code{brk_...} function.
}
