% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chop-by-width.R, R/breaks-by-width.R, R/tab.R
\name{chop_evenly}
\alias{chop_evenly}
\alias{brk_evenly}
\alias{tab_evenly}
\title{Chop into equal-width intervals}
\usage{
chop_evenly(x, intervals, ...)

brk_evenly(intervals)

tab_evenly(x, intervals, ...)
}
\arguments{
\item{x}{A vector.}

\item{intervals}{Integer: number of intervals to create.}

\item{...}{Passed to \code{\link[=chop]{chop()}}.}
}
\value{
\verb{chop_*} functions return a \code{\link[base:factor]{factor}} of the same length as \code{x}.

\verb{brk_*} functions return a \code{\link{function}} to create \code{breaks}.

\verb{tab_*} functions return a contingency \code{\link[base:table]{table}}.
}
\description{
\code{chop_evenly()} chops \code{x} into \code{intervals} intervals of equal width.
}
\examples{
chop_evenly(0:10, 5)

}
\seealso{
Other chopping functions:
\code{\link[=chop]{chop()}},
\code{\link[=chop_equally]{chop_equally()}},
\code{\link[=chop_fn]{chop_fn()}},
\code{\link[=chop_mean_sd]{chop_mean_sd()}},
\code{\link[=chop_n]{chop_n()}},
\code{\link[=chop_proportions]{chop_proportions()}},
\code{\link[=chop_quantiles]{chop_quantiles()}},
\code{\link[=chop_spikes]{chop_spikes()}},
\code{\link[=chop_width]{chop_width()}},
\code{\link[=fillet]{fillet()}}
}
\concept{chopping functions}
