% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chop.R
\name{chop_size}
\alias{chop_size}
\title{Chop into fixed-sized groups}
\usage{
chop_size(x, size, ...)
}
\arguments{
\item{x}{A numeric vector.}

\item{size}{Size of groups to return.}

\item{...}{Passed to \code{chop}.}
}
\description{
Chop into fixed-sized groups
}
\examples{
chop_size(1:10, 6)
}
