% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knife.R
\name{knife}
\alias{knife}
\title{Deprecated}
\usage{
knife(...)
}
\arguments{
\item{...}{Parameters for \code{\link[=chop]{chop()}}.}
}
\value{
A function.
}
\description{
\lifecycle{soft-deprecated}
\code{knife()} is deprecated in favour of \code{\link[purrr:partial]{purrr::partial()}}.
}
