% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kut.R
\name{kut}
\alias{kut}
\title{Cut numeric data into intervals}
\usage{
kut(x, breaks, labels, extend = NULL, drop = TRUE)
}
\arguments{
\item{x}{A numeric vector.}

\item{breaks}{See below.}

\item{labels}{See below.}

\item{extend}{Logical. Extend breaks to \code{+/-Inf}?}

\item{drop}{Logical. Drop unused levels from the result?}
}
\value{
A \link{factor} of the same length as \code{x}, representing the intervals containing
the value of \code{x}.
}
\description{
\code{kut} cuts \code{x} into intervals. It returns a factor of the same
length as \code{x}, representing which interval contains each element of \code{x}.
}
\details{
\code{breaks} may be a numeric vector, an object of class \link[=breaks-class]{breaks},
or a function.

If it is a vector, \code{breaks} gives the break endpoints.Repeated values create
singleton intervals. For example \code{breaks = c(1, 3, 3, 5)} creates 3
intervals: \code{[1, 3)}, \code{{3}} and \code{(3, 5]}. Default breaks are
left-closed except for the last interval; for right-closed breaks see
\code{\link[=brk_right]{brk_right()}}.

If \code{breaks} is a function it is called with a single argument, \code{x}, and
returns an object of class \code{breaks}.

\code{labels} may be a character vector. It should have the same length as the
number of intervals. Alternatively, use a \code{lbl_} function such as
\code{\link[=lbl_numerals]{lbl_numerals()}}.

If \code{extend} is \code{TRUE}, intervals will be extended to \code{[-Inf,
min(breaks))} and \code{(max(breaks), Inf]} (unless the first and last
endpoints are already \code{[-Inf} and \code{Inf]} respectively). If \code{extend} is \code{NULL}
(the default), intervals will be extended only if the data is outside their
range.

\code{NA} values in \code{x}, and values which are outside the (extended) endpoints,
return \code{NA}.

Note that \code{kut}, like all of R, uses binary arithmetic. Thus, numbers may
not be exactly equal to what you think they should be. There is an example
below.
}
\examples{
kut(1:3, 2)

kut(1:10, c(2, 5, 8))

kut(1:10, c(2, 5, 8), extend = FALSE)

kut(1:10, c(2, 5, 5, 8))

kut(1:10, brk_quantiles(c(0.25, 0.75)))

kut(1:10, c(2, 5, 8), labels = lbl_dash())

# floating point inaccuracy:
kut(0.3/3, c(0, 0.1, 0.1, 1))

}
\seealso{
cut

Other cutting functions: \code{\link{brk_mean_sd}},
  \code{\link{brk_n}}, \code{\link{brk_quantiles}},
  \code{\link{brk_width}}
}
\concept{cutting functions}
