% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breaks.R, R/kut.R
\name{brk_mean_sd}
\alias{brk_mean_sd}
\alias{kut_mean_sd}
\title{Cut by standard deviations}
\usage{
brk_mean_sd(sd = 3)

kut_mean_sd(x, sd = 3, ...)
}
\arguments{
\item{sd}{Whole number: include \code{sd} standard deviations on each side of
the mean.}

\item{x}{A numeric vector.}

\item{...}{Passed to \code{kut}.}
}
\value{
For  \code{kut_*} functions, a factor of the same length as \code{x}.
}
\description{
Cut by standard deviations
}
\examples{
kut_mean_sd(1:10)

kut(1:10, brk_mean_sd())

}
\seealso{
Other cutting functions: \code{\link{brk_n}},
  \code{\link{brk_quantiles}}, \code{\link{brk_width}},
  \code{\link{kut}}
}
\concept{cutting functions}
