% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breaks.R, R/kut.R
\name{brk_n}
\alias{brk_n}
\alias{kut_n}
\title{Cut into fixed-sized groups}
\usage{
brk_n(n)

kut_n(x, n, ...)
}
\arguments{
\item{n}{Integer: number of elements in each interval.}

\item{x}{A numeric vector.}

\item{...}{Passed to \code{kut}.}
}
\value{
For  \code{kut_*} functions, a factor of the same length as \code{x}.
}
\description{
\code{kut_n()} creates intervals containing a fixed number of elements. One
interval may have fewer elements.
}
\examples{
table(kut_n(1:10, 5))

table(kut_n(1:10, 4))

}
\seealso{
Other cutting functions: \code{\link{brk_mean_sd}},
  \code{\link{brk_quantiles}}, \code{\link{brk_width}},
  \code{\link{kut}}
}
\concept{cutting functions}
