% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breaks.R, R/kut.R
\name{brk_quantiles}
\alias{brk_quantiles}
\alias{kut_quantiles}
\alias{kut_deciles}
\alias{kut_equally}
\title{Cut by quantiles}
\usage{
brk_quantiles(probs, ...)

kut_quantiles(x, probs, ...)

kut_deciles(x, ...)

kut_equally(x, groups, ...)
}
\arguments{
\item{probs}{A vector of probabilities for the quantiles.}

\item{...}{Arguments passed to \code{\link[=quantile]{quantile()}}.}

\item{x}{A numeric vector.}

\item{groups}{Number of groups.}
}
\value{
For  \code{kut_*} functions, a factor of the same length as \code{x}.
}
\description{
\code{kut_quantiles} cuts data by quantiles. \code{kut_equally} cuts
data into equal-sized groups. \code{kut_deciles} is a convenience shortcut and
cuts into deciles.
}
\examples{
kut_quantiles(1:10, 1:3/4)

kut(1:10, brk_quantiles(1:3/4))

kut_deciles(1:10)

kut_equally(1:10, 5)

# to label by the quantiles themselves:
kut_quantiles(1:10, 1:3/4, lbl_intervals(raw = TRUE))

}
\seealso{
Other cutting functions: \code{\link{brk_mean_sd}},
  \code{\link{brk_n}}, \code{\link{brk_width}},
  \code{\link{kut}}
}
\concept{cutting functions}
