% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breaks.R, R/kut.R
\name{brk_width}
\alias{brk_width}
\alias{brk_evenly}
\alias{kut_width}
\alias{kut_evenly}
\title{Cut into equal-width intervals}
\usage{
brk_width(width, start)

brk_evenly(groups)

kut_width(x, width, start, ...)

kut_evenly(x, groups, ...)
}
\arguments{
\item{width}{Width of intervals.}

\item{start}{Leftpoint of first interval. By default the lowest finite \code{x}.}

\item{groups}{Integer: number of intervals to create.}

\item{x}{A numeric vector.}

\item{...}{Passed to \code{kut}.}
}
\value{
For  \code{kut_*} functions, a factor of the same length as \code{x}.
}
\description{
\code{kut_width()} cuts \code{x} into intervals of width \code{width}. \code{kut_evenly}
cuts \code{x} into \code{groups} intervals of equal width.
}
\examples{
kut_width(1:10, 2)

kut_width(1:10, 2, start = 0)

kut(1:10, brk_width(2, 0))

kut_evenly(1:10, 5)

}
\seealso{
Other cutting functions: \code{\link{brk_mean_sd}},
  \code{\link{brk_n}}, \code{\link{brk_quantiles}},
  \code{\link{kut}}
}
\concept{cutting functions}
