% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labels.R
\name{lbl_dashd}
\alias{lbl_dashd}
\title{Label chopped intervals like 1-3, 4-5, ...}
\usage{
lbl_dashd(
  symbol = em_dash(),
  raw = FALSE,
  fmt = NULL,
  first = NULL,
  last = NULL,
  ...
)
}
\arguments{
\item{raw}{Logical. Always use raw \code{breaks} in labels, rather than e.g. quantiles
or standard deviations?}

\item{fmt}{A format to be applied to the \code{breaks} used as labels. Can be a
string, passed into \code{\link[base:sprintf]{base::sprintf()}} or \code{\link[=format]{format()}} methods; or a
one-argument formatting function.}

\item{first}{Optional glue string for the first interval.}

\item{last}{Optional glue string for the last interval.}

\item{...}{Further arguments passed to \code{\link[glue:glue]{glue::glue()}}.}
}
\description{
Drop-in replacement for \code{\link[=lbl_dash]{lbl_dash()}}, but with all arguments except \code{symbol}
passed \code{\link[=lbl_glue]{lbl_glue()}}.
}
\examples{
chop(1:10, c(2, 5, 8), lbl_dashd())

chop(1:10, c(2, 5, 8), lbl_dashd(" to ", fmt = "\%.1f"))

chop(1:10, c(2, 5, 8), lbl_dashd(first = "< {r}"))

pretty <- function (x) prettyNum(x, big.mark = ",", digits = 1)
chop(runif(10) * 10000, c(3000, 7000), lbl_dash(" to ", fmt = pretty))
}
\seealso{
Other labelling functions: 
\code{\link{lbl_dash}()},
\code{\link{lbl_discrete}()},
\code{\link{lbl_format}()},
\code{\link{lbl_intervals}()},
\code{\link{lbl_manual}()},
\code{\link{lbl_seq}()}
}
\concept{labelling functions}
