% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labels-datetime.R
\name{lbl_date}
\alias{lbl_date}
\alias{lbl_datetime}
\title{Label dates and datetimes}
\usage{
lbl_date(
  fmt = "\%e \%b \%Y",
  symbol = "-",
  unit = as.difftime(1, units = "days"),
  single = "{l}",
  first = NULL,
  last = NULL
)

lbl_datetime(
  fmt = "\%H:\%M:\%S \%b \%e \%Y",
  symbol = "-",
  unit = NULL,
  single = "{l}",
  first = NULL,
  last = NULL
)
}
\arguments{
\item{fmt}{String, list or function. A format for break endpoints.}

\item{symbol}{String: separator to use for full ranges.}

\item{unit}{Optional interval unit for non-overlapping labels. If not \code{NULL},
.  endpoints are adjusted in the style of \code{\link[=lbl_discrete]{lbl_discrete()}}.}

\item{single}{Glue string: label for singleton intervals. See \code{\link[=lbl_glue]{lbl_glue()}}
for details. If \code{NULL}, singleton intervals will be labelled the same way
as other intervals.}

\item{first}{Glue string: override label for the first category. Write e.g.
\code{first = "<{r}"} to create a label like \code{"<18"}. See \code{\link[=lbl_glue]{lbl_glue()}}
for details.}

\item{last}{String: override label for the last category. Write e.g.
\code{last = ">{l}"} to create a label like \code{">65"}. See \code{\link[=lbl_glue]{lbl_glue()}}
for details.}
}
\value{
A function that creates a vector of labels.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

\code{lbl_date()} and \code{lbl_datetime()} produce nice labels for dates
and datetimes. Where possible ranges are simplified, like
like "13-14 Jul 2026" or "11:15-12:15 1 Dec 2025".
}
\section{Formatting endpoints}{


If \code{fmt} is not \code{NULL} then it is used to format the endpoints.
\itemize{
\item If \code{fmt} is a string, then numeric endpoints will be formatted by
\code{sprintf(fmt, breaks)}; other endpoints, e.g. \link[base:Dates]{Date} objects,
will be formatted by \code{format(breaks, fmt)}.
\item If \code{fmt} is a list, then it will be used as arguments to \link{format}.
\item If \code{fmt} is a function, it should take a vector of numbers (or other objects
that can be used as breaks) and return a character vector. It may be helpful
to use functions from the \code{{scales}} package, e.g. \code{\link[scales:label_comma]{scales::label_comma()}}.
}
}

\examples{
winter <- as.Date("2025-12-01") + 0:89
tab(winter, as.Date(c("2025-12-25", "2026-01-06")),
    labels = lbl_date())
new_year <- as.POSIXct("2025-12-31 23:00") + 0:120 * 60
round_midnight <- as.POSIXct(c("2025-12-31 23:59", "2026-01-01 00:05"))
tab(new_year, round_midnight,
    labels = lbl_datetime())
tab(new_year, round_midnight,
    labels = lbl_datetime(unit = as.difftime(1, units = "mins")))
}
\seealso{
Other labelling functions:
\code{\link[=lbl_dash]{lbl_dash()}},
\code{\link[=lbl_discrete]{lbl_discrete()}},
\code{\link[=lbl_endpoints]{lbl_endpoints()}},
\code{\link[=lbl_glue]{lbl_glue()}},
\code{\link[=lbl_intervals]{lbl_intervals()}},
\code{\link[=lbl_manual]{lbl_manual()}},
\code{\link[=lbl_midpoints]{lbl_midpoints()}},
\code{\link[=lbl_seq]{lbl_seq()}}
}
\concept{labelling functions}
