% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labels.R
\name{lbl_format}
\alias{lbl_format}
\title{Labels using breaks, with arbitrary formatting}
\usage{
lbl_format(fmt, fmt1 = "\%s")
}
\arguments{
\item{fmt}{A \code{\link[=sprintf]{sprintf()}}-style format.}

\item{fmt1}{Format for breaks consisting of a single value.}
}
\value{
A vector of labels for \code{chop}, or a function that creates labels.
}
\description{
Labels using breaks, with arbitrary formatting
}
\examples{
tab(1:10, c(1,3, 3, 7), label = lbl_format("\%s to \%s"))
tab(1:10, c(1,3, 3, 7), label = lbl_format("\%s to \%s", "Exactly \%s"))
}
