% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labels.R
\name{lbl_mean_sd}
\alias{lbl_mean_sd}
\title{Label standard deviations}
\usage{
lbl_mean_sd(sd)
}
\arguments{
\item{sd}{Number of standard deviations}
}
\value{
A vector of labels for \code{chop}, or a function that creates labels.
}
\description{
Label standard deviations
}
