% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labels-single.R
\name{lbl_midpoints}
\alias{lbl_midpoints}
\title{Label chopped intervals by their midpoints}
\usage{
lbl_midpoints(
  fmt = NULL,
  single = NULL,
  first = NULL,
  last = NULL,
  raw = deprecated()
)
}
\arguments{
\item{fmt}{String, list or function. A format for break endpoints.}

\item{single}{Glue string: label for singleton intervals. See \code{\link[=lbl_glue]{lbl_glue()}}
for details. If \code{NULL}, singleton intervals will be labelled the same way
as other intervals.}

\item{first}{Glue string: override label for the first category. Write e.g.
\code{first = "<{r}"} to create a label like \code{"<18"}. See \code{\link[=lbl_glue]{lbl_glue()}}
for details.}

\item{last}{String: override label for the last category. Write e.g.
\code{last = ">{l}"} to create a label like \code{">65"}. See \code{\link[=lbl_glue]{lbl_glue()}}
for details.}

\item{raw}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}. Throws an error.
Use the \code{raw} argument to \code{\link[=chop]{chop()}} instead.}
}
\value{
A function that creates a vector of labels.
}
\description{
This uses the midpoint of each interval for
its label.
}
\section{Formatting endpoints}{


If \code{fmt} is not \code{NULL} then it is used to format the endpoints.
\itemize{
\item If \code{fmt} is a string, then numeric endpoints will be formatted by
\code{sprintf(fmt, breaks)}; other endpoints, e.g. \link[base:Dates]{Date} objects,
will be formatted by \code{format(breaks, fmt)}.
\item If \code{fmt} is a list, then it will be used as arguments to \link{format}.
\item If \code{fmt} is a function, it should take a vector of numbers (or other objects
that can be used as breaks) and return a character vector. It may be helpful
to use functions from the \code{{scales}} package, e.g. \code{\link[scales:label_comma]{scales::label_comma()}}.
}
}

\examples{
chop(1:10, c(2, 5, 8), lbl_midpoints())
}
\seealso{
Other labelling functions:
\code{\link[=lbl_dash]{lbl_dash()}},
\code{\link[=lbl_date]{lbl_date()}},
\code{\link[=lbl_discrete]{lbl_discrete()}},
\code{\link[=lbl_endpoints]{lbl_endpoints()}},
\code{\link[=lbl_glue]{lbl_glue()}},
\code{\link[=lbl_intervals]{lbl_intervals()}},
\code{\link[=lbl_manual]{lbl_manual()}},
\code{\link[=lbl_seq]{lbl_seq()}}
}
\concept{labelling functions}
