% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labels.R
\name{lbl_quantiles}
\alias{lbl_quantiles}
\title{Labels suitable for quantiles}
\usage{
lbl_quantiles(quantiles)
}
\arguments{
\item{quantiles}{A vector of quantiles.}
}
\value{
A vector of labels for \code{chop}, or a function that creates labels.
}
\description{
Labels suitable for quantiles
}
\examples{
lbl_quantiles(c(0.25, 0.5, 0.75))
}
