% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/santoku-cast.R
\name{santoku-cast}
\alias{santoku-cast}
\alias{santoku_cast_common.default}
\alias{santoku_cast_common.double}
\alias{santoku_cast_common.Date}
\alias{santoku_cast_common.POSIXct}
\alias{santoku_cast_common.ts}
\alias{santoku_cast_common.zoo}
\alias{santoku_cast_common.integer64}
\alias{santoku_cast_common.hexmode}
\alias{santoku_cast_common.octmode}
\title{Internal functions}
\usage{
santoku_cast_common.default(x, y)

santoku_cast_common.double(x, y)

santoku_cast_common.Date(x, y)

santoku_cast_common.POSIXct(x, y)

santoku_cast_common.ts(x, y)

santoku_cast_common.zoo(x, y)

santoku_cast_common.integer64(x, y)

santoku_cast_common.hexmode(x, y)

santoku_cast_common.octmode(x, y)
}
\arguments{
\item{x, y}{Vectors to cast.}
}
\value{
A list.
}
\description{
Internal functions
}
\keyword{Do}
\keyword{These}
\keyword{are}
\keyword{functions.}
\keyword{internal}
\keyword{not}
\keyword{use.}
