% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab.R
\name{tab}
\alias{tab}
\alias{tab_width}
\alias{tab_evenly}
\alias{tab_n}
\alias{tab_mean_sd}
\title{Tabulate data by intervals}
\usage{
tab(...)

tab_width(...)

tab_evenly(...)

tab_n(...)

tab_mean_sd(...)
}
\arguments{
\item{...}{Passed to \code{chop}}
}
\value{
A \code{\link[=table]{table()}}.
}
\description{
These functions call their related \code{chop_xxx} function, and call \code{\link[=table]{table()}}
on the result.
}
\examples{
tab(1:10, c(2, 5, 8))

tab_mean_sd(1:10)
}
