% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbar.R
\name{tbar}
\alias{tbar}
\alias{tbar.default}
\alias{tbar.table}
\title{Print a terminal barplot of data}
\usage{
tbar(x, ..., width = 60)

\method{tbar}{default}(x, ..., width = 60, useNA = "ifany")

\method{tbar}{table}(x, ..., width = 60)
}
\arguments{
\item{x}{A vector or table.}

\item{...}{Not currently used.}

\item{width}{Maximum width of the barplot in characters.}

\item{useNA}{Passed to \code{\link[=table]{table()}}: one of \code{"no"}, \code{"ifany"} or \code{"always"}.}
}
\value{
\code{x} or \code{table(x)}, invisibly.
}
\description{
Print a terminal barplot of data
}
\examples{
tbar(ChickWeight$Diet)
props <- proportions(table(ChickWeight$Diet))
tbar(props)
tbar(chop(rnorm(100), -3:3))
}
