% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breaks.R
\name{brk-left-right}
\alias{brk-left-right}
\alias{brk_left}
\alias{brk_right}
\title{Left- or right-closed breaks}
\usage{
brk_left(breaks)

brk_right(breaks)
}
\arguments{
\item{breaks}{A numeric vector.}
}
\value{
A (function which returns an) object of class \code{breaks}.
}
\description{
\lifecycle{questioning}
}
\details{
These functions are in the "questioning" stage because they clash with the
\code{left} argument to \code{\link[=chop]{chop()}} and friends.

These functions override the \code{left} argument of \code{\link[=chop]{chop()}}.
}
\examples{
chop(5:7, brk_left(5:7))

chop(5:7, brk_right(5:7))

chop(5:7, brk_left(5:7))

}
