% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breaks.R
\name{brk_mean_sd}
\alias{brk_mean_sd}
\title{Title}
\usage{
brk_mean_sd(sd = 3)
}
\arguments{
\item{sd}{Whole number: include \code{sd} standard deviations on each side of
the mean.}
}
\value{
A (function which returns an) object of class \code{breaks}.
}
\description{
Title
}
\examples{
tab(rnorm(20), brk_mean_sd())
}
