% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breaks.R
\name{brk_width}
\alias{brk_width}
\alias{brk_evenly}
\title{Equal-width breaks}
\usage{
brk_width(width, start)

brk_evenly(groups)
}
\arguments{
\item{width}{Width of intervals.}

\item{start}{Leftpoint of first interval. By default the lowest finite \code{x}.}

\item{groups}{Integer: number of intervals to create.}
}
\value{
A (function which returns an) object of class \code{breaks}.
}
\description{
\code{brk_width()} creates intervals of width \code{width}. \code{brk_evenly()} creates
\code{groups} intervals of equal width.
}
\examples{
chop(runif(10), brk_width(0.2, 0))
}
