% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chop.R
\name{chop}
\alias{chop}
\title{Cut numeric data into intervals}
\usage{
chop(x, breaks, labels = lbl_intervals(), extend = TRUE, drop = TRUE)
}
\arguments{
\item{x}{A numeric vector.}

\item{breaks}{See below.}

\item{labels}{See below.}

\item{extend}{Logical. Extend breaks to \code{+/-Inf}?}

\item{drop}{Logical. Drop unused levels from the result?}
}
\value{
A \link{factor}, or an integer if \code{labels} is \code{NULL}, of the same length as \code{x},
representing the intervals containing the value of \code{x}.
}
\description{
\code{chop} cuts \code{x} into intervals. It returns a factor of the same
length as \code{x}, representing which interval contains each element of \code{x}.
}
\details{
\code{breaks} may be a numeric vector, an object of class \link[=breaks-class]{breaks},
or a function.

If it is a vector, \code{breaks} gives the break endpoints.Repeated values create
singleton intervals. For example \code{breaks = c(1, 3, 3, 5)} creates 3
intervals: \code{[1, 3)}, \code{{3}} and \code{(3, 5]}. Default breaks are
left-closed except the end; for right-closed breaks see \code{\link[=brk_right]{brk_right()}}.

If \code{breaks} is a function it is called with a single argument, \code{x}, and
returns an object of class \code{breaks}.

\code{labels} may be a character vector or a function. If it is a character
vector it should have the same length as the number of intervals - i.e.
one more than \code{length(breaks)} if \code{extend} is \code{TRUE}, one less otherwise.

If \code{labels} is a function it should take an object of class \code{breaks} and
return an appropriate character vector.

If \code{extend} is \code{TRUE}, intervals will be extended to \code{[-Inf,
min(breaks))} and \code{(max(breaks), Inf]}, unless those endpoints are
already infinite.

\code{NA} values in \code{x}, and values which are outside the (extendeD) endpoints,
return \code{NA}.
}
\examples{
chop(1:3, 2)
chop(rnorm(10), -2:2)
chop(rnorm(10), -2:2, extend = FALSE)
chop(rnorm(10), brk_quantiles(c(0.25, 0.75)))
chop(rnorm(10), -2:2, labels = lbl_dash())
}
\seealso{
cut
}
