% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breaks.R, R/chop.R
\name{brk_mean_sd}
\alias{brk_mean_sd}
\alias{chop_mean_sd}
\title{Chop by standard deviations}
\usage{
brk_mean_sd(sd = 3)

chop_mean_sd(x, sd = 3, ...)
}
\arguments{
\item{sd}{Whole number: include \code{sd} standard deviations on each side of
the mean.}

\item{x}{A numeric vector.}

\item{...}{Passed to \code{chop}.}
}
\value{
For  \verb{chop_*} functions, a factor of the same length as \code{x}.
}
\description{
Chop by standard deviations
}
\examples{
chop_mean_sd(1:10)

chop(1:10, brk_mean_sd())

}
\seealso{
Other chopping functions: 
\code{\link{brk_n}()},
\code{\link{brk_quantiles}()},
\code{\link{brk_width}()},
\code{\link{chop}()},
\code{\link{fillet}()}
}
\concept{chopping functions}
