% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chop.R, R/breaks.R, R/tab.R
\name{chop_mean_sd}
\alias{chop_mean_sd}
\alias{brk_mean_sd}
\alias{tab_mean_sd}
\title{Chop by standard deviations}
\usage{
chop_mean_sd(x, sd = 3, ...)

brk_mean_sd(sd = 3)

tab_mean_sd(x, sd = 3, ...)
}
\arguments{
\item{x}{A vector.}

\item{sd}{Positive number: include up to \code{sd} standard deviations.}

\item{...}{Passed to \code{\link[=chop]{chop()}}.}
}
\value{
\verb{chop_*} functions return a \code{\link{factor}} of the same length as \code{x}.

\verb{brk_*} functions return a \code{\link{function}} to create \code{breaks}.

\verb{tab_*} functions return a contingency \code{\link[=table]{table()}}.
}
\description{
Intervals of width 1 standard deviation are included on either side of the
mean.
The outermost pair of intervals will be shorter if \code{sd} is not a whole
number.
}
\examples{
chop_mean_sd(1:10)

chop(1:10, brk_mean_sd())

tab_mean_sd(1:10)

}
\seealso{
Other chopping functions: 
\code{\link{chop_n}()},
\code{\link{chop_quantiles}()},
\code{\link{chop_width}()},
\code{\link{fillet}()},
\code{\link{tab}()}
}
\concept{chopping functions}
