% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chop.R
\name{chop_mean_sd}
\alias{chop_mean_sd}
\title{Chop by standard deviations}
\usage{
chop_mean_sd(x, sd = 3, labels = lbl_mean_sd(sd), ...)
}
\arguments{
\item{x}{A numeric vector.}

\item{sd}{How many standard deviations to include on each side of the mean}

\item{...}{Passed to \code{chop}.}
}
\description{
Chop by standard deviations
}
\examples{
chop_mean_sd(rnorm(20))
}
