% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chop.R
\name{chop_n}
\alias{chop_n}
\title{Chop into fixed-sized groups}
\usage{
chop_n(x, n, ...)
}
\arguments{
\item{x}{A numeric vector.}

\item{n}{Size of groups to return.}

\item{...}{Passed to \code{chop}.}
}
\description{
Chop into fixed-sized groups
}
\examples{
chop_n(1:10, 6)
}
