% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chop.R
\name{chop_quantiles}
\alias{chop_quantiles}
\alias{chop_deciles}
\alias{chop_equally}
\title{Chop by quantiles.}
\usage{
chop_quantiles(x, probs, ...)

chop_deciles(x, ...)

chop_equally(x, groups, ...)
}
\arguments{
\item{x}{A numeric vector.}

\item{probs}{A vector of probabilities for the quantiles.}

\item{...}{Passed to \code{chop}.}

\item{groups}{Number of groups}
}
\description{
\code{chop_quantiles} chops data by quantiles. \code{chop_equally} chops
data into equal-sized groups. \code{chop_deciles} is a convenience shortcut and
chops into deciles.
}
\examples{
chop_quantiles(rnorm(10), c(1/3, 2/3))
chop_equally(rnorm(10), 5)
}
