% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chop.R, R/breaks.R
\name{chop_quantiles}
\alias{chop_quantiles}
\alias{chop_deciles}
\alias{chop_equally}
\alias{brk_quantiles}
\title{Chop by quantiles}
\usage{
chop_quantiles(x, probs, ...)

chop_deciles(x, ...)

chop_equally(x, groups, ...)

brk_quantiles(probs, ...)
}
\arguments{
\item{x}{A numeric vector.}

\item{probs}{A vector of probabilities for the quantiles.}

\item{...}{Arguments passed to \code{\link[=quantile]{quantile()}}.}

\item{groups}{Number of groups.}
}
\value{
For  \verb{chop_*} functions, a factor of the same length as \code{x}.
}
\description{
\code{chop_quantiles} chops data by quantiles. \code{chop_equally} chops
data into equal-sized groups. \code{chop_deciles} is a convenience shortcut and
chops into deciles.
}
\examples{
chop_quantiles(1:10, 1:3/4)

chop(1:10, brk_quantiles(1:3/4))

chop_deciles(1:10)

chop_equally(1:10, 5)

# to label by the quantiles themselves:
chop_quantiles(1:10, 1:3/4, lbl_intervals(raw = TRUE))

}
\seealso{
Other chopping functions: 
\code{\link{chop_mean_sd}()},
\code{\link{chop_n}()},
\code{\link{chop_width}()},
\code{\link{chop}()},
\code{\link{fillet}()}
}
\concept{chopping functions}
