% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chop.R
\name{chop_quantiles}
\alias{chop_quantiles}
\alias{chop_equal}
\title{Chop by quantiles.}
\usage{
chop_quantiles(x, quantiles, labels = lbl_quantiles(quantiles), ...)

chop_equal(x, n, ...)
}
\arguments{
\item{x}{A numeric vector.}

\item{quantiles}{A vector of quantiles. You don't need to include 0 or 1.}

\item{labels}{See \code{\link[=chop]{chop()}}.}

\item{...}{Passed to \code{chop}.}

\item{n}{Number of groups}
}
\description{
\code{chop_quantiles} chops data by quantiles. \code{chop_equal} chops
data into equal-sized groups.
}
\examples{
chop_quantiles(rnorm(10), c(1/3, 2/3))
chop_equal(rnorm(10), 5)
}
