% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chop.R
\name{chop_width}
\alias{chop_width}
\title{Chop into equal-width intervals}
\usage{
chop_width(x, width, start, ...)
}
\arguments{
\item{x}{A numeric vector.}

\item{width}{Width of intervals.}

\item{start}{Leftpoint of first interval. By default the lowest finite \code{x}.}

\item{...}{Passed to \code{chop}.}
}
\description{
\code{chop_width()} chops \code{x} into intervals of equal width.
}
\examples{
chop_width(rnorm(10), 2)
chop_width(0:10, 2)
}
