% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{exactly}
\alias{exactly}
\title{Syntactic sugar}
\usage{
exactly(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
\code{rep(x, each = 2)}.
}
\description{
\code{exactly} lets you write \code{chop(x, c(1, exactly(2), 3))}. This
is the same as \code{chop(x, c(1, 2, 2, 3))} but conveys your intent more
clearly.
}
\examples{
tab(1:5, exactly(3))
}
