% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knife.R
\name{knife}
\alias{knife}
\title{Create a cutting function}
\usage{
knife(...)
}
\arguments{
\item{...}{Parameters for \code{\link[=kut]{kut()}}.}
}
\description{
\code{knife()} returns a one-argument function for cutting data.
}
\examples{
kut_six <- knife(breaks = -2:2, labels = lbl_LETTERS())
kut_six(rnorm(10))
kut_six(rnorm(10))
}
