% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knife.R
\name{knife}
\alias{knife}
\title{Create a chopping function}
\usage{
knife(...)
}
\arguments{
\item{...}{Parameters for \code{\link[=chop]{chop()}}.}
}
\description{
\code{knife()} returns a one-argument function for chopping data.
}
\examples{
chop_six <- knife(breaks = -2:2, labels = letters[1:6])
chop_six(rnorm(10))
chop_six(rnorm(10))
}
