% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labels.R
\name{lbl_dash}
\alias{lbl_dash}
\title{Labels like 1 - 3, 4 - 5, ...}
\usage{
lbl_dash(symbol = " - ")
}
\arguments{
\item{symbol}{String: symbol to use for the dash.}
}
\value{
A vector of labels for \code{chop}, or a function that creates labels.
}
\description{
Labels like 1 - 3, 4 - 5, ...
}
\examples{
tab(0:20, c(0, 5, 5, 10), lbl_dash())
}
