% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labels.R
\name{lbl_dash}
\alias{lbl_dash}
\title{Labels like 1 - 3, 4 - 5, ...}
\usage{
lbl_dash(symbol = " - ", raw = FALSE)
}
\arguments{
\item{symbol}{String: symbol to use for the dash.}

\item{raw}{Logical. Always use raw \code{breaks} in labels, rather than e.g. quantiles
or standard deviations.}
}
\value{
A vector of labels for \code{chop}, or a function that creates labels.
}
\description{
Labels like 1 - 3, 4 - 5, ...
}
\examples{
chop(1:10, c(2, 5, 8), lbl_dash())

chop(1:10, c(2, 5, 8), lbl_dash(" to "))

}
\seealso{
Other labelling functions: 
\code{\link{lbl_format}()},
\code{\link{lbl_integer}()},
\code{\link{lbl_intervals}()},
\code{\link{lbl_manual}()},
\code{\link{lbl_seq}()}
}
\concept{labelling functions}
