% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labels.R
\name{lbl_dash}
\alias{lbl_dash}
\title{Labels like 1 - 3, 4 - 5, ...}
\usage{
lbl_dash(symbol = " - ", raw = FALSE, fmt = NULL)
}
\arguments{
\item{symbol}{String: symbol to use for the dash.}

\item{raw}{Logical. Always use raw \code{breaks} in labels, rather than e.g. quantiles
or standard deviations?}

\item{fmt}{A format string, passed into \code{\link[base:sprintf]{base::sprintf()}}, or to \code{\link[=format]{format()}} methods
for non-numeric breaks. Default varies according to the type of breaks.}
}
\value{
A vector of labels for \code{chop}, or a function that creates labels.
}
\description{
This label style is user-friendly, but doesn't distinguish between
left- and right-closed intervals.
}
\examples{
chop(1:10, c(2, 5, 8), lbl_dash())

chop(1:10, c(2, 5, 8), lbl_dash(" to ", fmt = "\%.1f"))

}
\seealso{
Other labelling functions: 
\code{\link{lbl_discrete}()},
\code{\link{lbl_format}()},
\code{\link{lbl_intervals}()},
\code{\link{lbl_manual}()},
\code{\link{lbl_seq}()}
}
\concept{labelling functions}
