% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labels.R
\name{lbl_date}
\alias{lbl_date}
\title{Labels for date or date-time objects}
\usage{
lbl_date(fmt = "\%Y-\%m-\%d", symbol = " to ")
}
\arguments{
\item{fmt}{Format, to be passed to \code{\link[=format.POSIXct]{format.POSIXct()}}.}

\item{symbol}{String to place between dates.}
}
\value{
A vector of labels for \code{chop}, or a function that creates labels.
}
\description{
Labels for date or date-time objects
}
\seealso{
Other labelling functions: 
\code{\link{lbl_dash}()},
\code{\link{lbl_format}()},
\code{\link{lbl_integer}()},
\code{\link{lbl_intervals}()},
\code{\link{lbl_manual}()},
\code{\link{lbl_seq}()}
}
\concept{labelling functions}
