% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labels.R
\name{lbl_discrete}
\alias{lbl_discrete}
\title{Label discrete data}
\usage{
lbl_discrete(
  symbol = em_dash(),
  unit = 1,
  fmt = NULL,
  first = NULL,
  last = NULL
)
}
\arguments{
\item{symbol}{String: symbol to use for the dash.}

\item{unit}{Minimum difference between distinct values of data.
For integers, 1.}

\item{fmt}{A format to be applied to the \code{breaks} used as labels. Can be a
string, passed into \code{\link[=sprintf]{sprintf()}} (numbers) or \code{\link[=format]{format()}} (dates, times);
or a one-argument formatting function.}

\item{first}{String: override label for the first category. Passed to
\code{\link[glue:glue]{glue::glue()}}: write e.g. \code{first = "< {r}"} to create a label like \code{"< 18"}.}

\item{last}{String: override label for the last category. Passed to
\code{\link[glue:glue]{glue::glue()}}: write e.g. \verb{last = "> \{l\}} to create a label like \code{"> 65"}.}
}
\value{
A vector of labels for \code{chop}, or a function that creates labels.
}
\description{
\lifecycle{experimental}
}
\details{
\code{lbl_discrete} creates labels for discrete data such as integers.
For example, breaks
\code{c(1, 3, 4, 6, 7)} are labelled: \verb{"1 - 2", "3", "4 - 5", "6 - 7"}.

No check is done that the data is discrete-valued. If it isn't, then
these labels may be misleading. Here, discrete-valued means that if
\code{x < y}, then \code{x <= y - unit}.

Be aware that Date objects may have non-integer values. See \link{Date}.
}
\examples{
tab(1:7, c(1, 3, 5), lbl_discrete())

tab(1:7, c(3, 5), lbl_discrete(first = "<= \%s"))

# Misleading labels for non-integer data
chop(2.5, c(1, 3, 5), lbl_discrete())

tab(1:7 * 1000, c(1, 3, 5) * 1000, lbl_discrete(unit = 1000))
}
\seealso{
Other labelling functions: 
\code{\link{lbl_dash}()},
\code{\link{lbl_glue}()},
\code{\link{lbl_intervals}()},
\code{\link{lbl_manual}()},
\code{\link{lbl_seq}()}
}
\concept{labelling functions}
