% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labels.R
\name{lbl_format}
\alias{lbl_format}
\title{Labels with arbitrary formatting}
\usage{
lbl_format(fmt, fmt1 = "\%.3g", raw = FALSE)
}
\arguments{
\item{fmt}{A \code{\link[=sprintf]{sprintf()}}-style format. Default varies according to the
type of breaks.}

\item{fmt1}{Format for breaks consisting of a single value.}

\item{raw}{Logical. Always use raw \code{breaks} in labels, rather than e.g. quantiles
or standard deviations?}
}
\value{
A vector of labels for \code{chop}, or a function that creates labels.
}
\description{
These labels use \code{\link[base:sprintf]{base::sprintf()}} to format numeric breaks.
}
\examples{
tab(1:10, c(1,3, 3, 7), label = lbl_format("\%.3g to \%.3g"))
tab(1:10, c(1,3, 3, 7), label = lbl_format("\%.3g to \%.3g", "Exactly \%.3g"))
tab(rnorm(10), seq(-2, 2, .5), label = lbl_format("\%.2f to \%.2f"))
}
\seealso{
Other labelling functions: 
\code{\link{lbl_dash}()},
\code{\link{lbl_integer}()},
\code{\link{lbl_intervals}()},
\code{\link{lbl_manual}()},
\code{\link{lbl_seq}()}
}
\concept{labelling functions}
