% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labels.R
\name{lbl_glue}
\alias{lbl_glue}
\title{Label chopped intervals using the {glue} package}
\usage{
lbl_glue(
  label,
  fmt = NULL,
  single = NULL,
  first = NULL,
  last = NULL,
  raw = FALSE,
  ...
)
}
\arguments{
\item{label}{A glue string passed to \code{\link[glue:glue]{glue::glue()}}.}

\item{fmt}{A format to be applied to the break endpoints used in labels. \code{fmt}
can be a string, passed into \code{\link[=sprintf]{sprintf()}} (numbers) or \code{\link[=format]{format()}} (dates, times);
or a one-argument formatting function.}

\item{single}{Glue string: label for singleton intervals. See \code{\link[=lbl_glue]{lbl_glue()}}
for details.}

\item{first}{Glue string: override label for the first category. Write e.g.
\code{first = "< {r}"} to create a label like \code{"< 18"}. See \code{\link[=lbl_glue]{lbl_glue()}}
for details.}

\item{last}{String: override label for the last category. Write e.g.
\verb{last = "> \{l\}} to create a label like \code{"> 65"}. See \code{\link[=lbl_glue]{lbl_glue()}}
for details.}

\item{raw}{Logical. Always use raw \code{breaks} in labels, rather than e.g. quantiles
or standard deviations?}

\item{...}{Further arguments passed to \code{\link[glue:glue]{glue::glue()}}.}
}
\value{
A vector of labels for \code{chop}, or a function that creates labels.
}
\description{
Use \code{"{l}"} and \code{"{r}"} to show the left and right endpoints of the intervals.
}
\details{
Variables \code{l_closed} and \code{r_closed} are \code{TRUE} if the left and right
endpoints are closed.

The endpoints themselves are formatted using \code{fmt} before being passed to
\code{glue()}.
}
\examples{
tab(1:10, c(1, 3, 3, 7),
    label = lbl_glue("{l} to {r}", single = "Exactly {l}"))

tab(1:10 * 1000, c(1, 3, 5, 7) * 1000,
    label = lbl_glue("{l}-{r}", fmt = \(x) prettyNum(x, big.mark=',')))

# reproducing lbl_intervals():
glue_string <- "{ifelse(l_closed, '[', '(')}{l}, {r}{ifelse(r_closed, ']', ')')}"
tab(1:10, c(1, 3, 3, 7), label = lbl_glue(glue_string, single = "{{{l}}}"))

}
\seealso{
Other labelling functions: 
\code{\link{lbl_dash}()},
\code{\link{lbl_discrete}()},
\code{\link{lbl_intervals}()},
\code{\link{lbl_manual}()},
\code{\link{lbl_seq}()}
}
\concept{labelling functions}
