% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labels.R
\name{lbl_glue}
\alias{lbl_glue}
\title{Label chopped intervals using the \code{glue} package}
\usage{
lbl_glue(fmt, fmt1 = fmt, first = NULL, last = NULL, raw = FALSE, ...)
}
\arguments{
\item{fmt}{A glue string passed to \code{\link[glue:glue]{glue::glue()}}.}

\item{fmt1}{Optional glue string for singleton intervals.}

\item{first}{Optional glue string for the first interval.}

\item{last}{Optional glue string for the last interval.}

\item{raw}{Logical. Always use raw \code{breaks} in labels, rather than e.g. quantiles
or standard deviations?}

\item{...}{Further arguments passed to \code{\link[glue:glue]{glue::glue()}}.}
}
\value{
A vector of labels for \code{chop}, or a function that creates labels.
}
\description{
\lifecycle{experimental}
}
\details{
In glue strings, variables \code{l} and \code{r} represent the left and right
endpoints. Variables \code{l_closed} and \code{r_closed} are \code{TRUE} if the left
and right endpoints are closed respectively.
}
\examples{
if (requireNamespace("glue")) {
  tab(1:10, c(1,3, 3, 7),
      label = lbl_glue("{l} to {r}", fmt1 = "Exactly {l}"))

 tab(1:10 * 1000, c(1,3, 5, 7) * 1000, label =
   lbl_glue("{prettyNum(l, big.mark=',')}-{prettyNum(r, big.mark=',')}"))

  # reproducing lbl_intervals():
  glue_string <- "{ifelse(l_closed, '[', '(')}{l}, {r}{ifelse(r_closed, ']', ')')}"
  tab(1:10, c(1,3, 3, 7), label = lbl_glue(glue_string, fmt1 = "{{{l}}}"))


}

}
