% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labels.R
\name{lbl_integer}
\alias{lbl_integer}
\title{Labels for integer data}
\usage{
lbl_integer(symbol = " - ")
}
\arguments{
\item{symbol}{String: symbol to use for the dash.}
}
\value{
A vector of labels for \code{chop}, or a function that creates labels.
}
\description{
\code{lbl_integer} creates labels for integer data. For example, breaks
\code{c(1, 3, 4, 6, 7)} get labels \verb{"1 - 2", "3", "4 - 5", "6 - 7"}.
}
\details{
No check is done that the data is integer-valued. If it isn't, then
these labels may be misleading.
}
\examples{
tab(1:7, c(1, 3, 5), lbl_integer())

# Misleading labels for non=integer data
chop(2.5, c(1, 3, 5), lbl_integer())
}
\seealso{
Other labelling functions: 
\code{\link{lbl_dash}()},
\code{\link{lbl_date}()},
\code{\link{lbl_format}()},
\code{\link{lbl_intervals}()},
\code{\link{lbl_manual}()},
\code{\link{lbl_seq}()}
}
\concept{labelling functions}
