% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labels.R
\name{lbl_intervals}
\alias{lbl_intervals}
\title{Labels using set notation}
\usage{
lbl_intervals(fmt = "\%.2g")
}
\arguments{
\item{fmt}{A \code{\link[=sprintf]{sprintf()}}-style format.}
}
\value{
A vector of labels for \code{chop}, or a function that creates labels.
}
\description{
Labels using set notation
}
\details{
\code{fmt} is applied to interval endpoints.
}
\examples{
tab(rnorm(100), c(-3, 0, 0, 3), labels = lbl_intervals())
tab(rnorm(100), c(-3, 0, 0, 3), labels = lbl_intervals("\%.2f"))
}
