% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labels.R
\name{sequence-labels}
\alias{sequence-labels}
\alias{lbl_numerals}
\alias{lbl_roman}
\alias{lbl_ROMAN}
\alias{lbl_letters}
\alias{lbl_LETTERS}
\alias{lbl_sequence}
\title{Label in sequence}
\usage{
lbl_numerals(fmt = "\%s")

lbl_roman(fmt = "\%s")

lbl_ROMAN(fmt = "\%s")

lbl_letters(fmt = "\%s")

lbl_LETTERS(fmt = "\%s")

lbl_sequence(sequence, fmt = "\%s")
}
\arguments{
\item{fmt}{A \code{\link[=sprintf]{sprintf()}}-style format.}

\item{sequence}{A character vector.}
}
\value{
A vector of labels for \code{chop}, or a function that creates labels.
}
\description{
These functions label intervals sequentially from lowest to highest.
}
\details{
\code{lbl_numerals()} uses arabic numerals.

\code{lbl_letters()} and \code{lbl_LETTERS()}
uses lower-case and upper-case letters respectively.

\code{lbl_roman()} and
\code{lbl_ROMAN()} use lower-case and upper-case Roman numerals respectively.

\code{lbl_sequence()} uses an arbitrary sequence. If the sequence is shorter than
the number of breaks, it will be pasted with itself and repeated
as necessary.
}
\examples{

tab(1:10, c(3, 4), lbl_numerals())
tab(1:10, c(3, 4), lbl_numerals("(\%s)"))
tab(1:10, c(3, 4), lbl_letters())
tab(1:10, c(3, 4), lbl_LETTERS())
tab(1:10, c(3, 4), lbl_roman())
tab(1:10, c(3, 4), lbl_ROMAN())
tab(1:10, c(3, 4), lbl_sequence(c("x", "y", "z")))

# if labels need repeating:
tab(1:10, 1:10, lbl_sequence(c("x", "y", "z")))
}
