% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab.R
\name{tab}
\alias{tab}
\alias{tab_width}
\alias{tab_size}
\title{Tabulate data by intervals}
\usage{
tab(...)

tab_width(...)

tab_size(...)
}
\arguments{
\item{drop, ...}{Passed to \code{chop}}
}
\value{
A \code{\link[=table]{table()}}.
}
\description{
These functions call \code{\link[=chop]{chop()}} and friends, and call \code{\link[=table]{table()}}
on the result.
}
\examples{
tab(rnorm(100), -2:2)
}
